Save syndication links into the post frontmatter (#13)
* add constants for common date formats

* Frontmatter handling updates

- parse the whole thing into a map[string]string
- get title/slug/date from parsed frontmatter

* update tests

* Add post options and new constructor

- this is to make sure that frontmatter is initialized as a map[string]string properly

* update sql code to account for new fields, field order

* remove extraneous comment

* clean up parser options to match the goldfrog service

* update test data

* THIS BRANCH IS SO SCREWED

* quick commit it while it compiles

* switch out the weird imported structs

for IDENTICAL local ones that still work (learned something new about struct equivalence)

* Just default the new post data to now()

* Work on passing syndicator options through better

* move all syndication code into it's own package

it doesn't clean up the syndication process a HUGE amount, but it's much more
standardized. I worked out how to capture the ids and links from the
syndication providers so we can give that data to the templates

* don't use goroutines to post mentions, it doesn't save *that* much.

* implement new syndication process

Also solved the "save the post again" problem. Without the ID attached,
SavePost could not save the new data. Solved it by loading the post
from the slug before doing the syndication, then save it out again with
the links added to the frontmatter